Add decode_timestamp and decode_randomness.
For now, these implement all of the unrolled loop logic themselves
to save on performance for additional function calls and bytes
slice 'decode_ulid' would have to make if this was refactored
properly.

I will come back to this consideration after running benchmarks to
see if the code duplication is worth the performance.