Fix api.from_int bit_length check.
Previously, this would throw a ValueError for any integer that
didn't require 16 bytes to store. This worked for the most part
until the timestamp part was low enough to be zero padded. Once
this happened, we technically only need 15 bytes to store the
integer, so an exception would be incorrectly thrown.

With this change, a ValueError will only be raised when an overflow
would occur. If we need less than 16 bytes to store the integer,
the conversion to bytes will zero pad it and the equivalent value
will be created.