Add 'parse' to public API
The 'parse' method should take a value of any supported type
and make a best guess attempt to convert it to a ULID instance.

Callers should only use this method when they're confident that they've
received a value that can be represented as a ULID but are unsure what
format/type it is in.

Callers should use the other from_* methods when the type/format is known.