Add microsecond provider
The microsecond provider uses more precise clock resolution to
create randomness values that should monotonically increase on
timestamp (millisecond) collision.

When creating a new ULID, we extract the number of microseconds from
our timestamp and use them as the first two bytes of the randomness value.

Caveats:

When using `ulid.from_timestamp`, we fallback to the default implementation
of 10 bytes of pure randomness. The reason for this, is that the timestamp
that is being passed in by the caller is in milliseconds. Since we don't have
a more accurate timestamp to "steal" bytes from, there's not much we can do. We
cannot make a new epoch timestamp as the microsecond remainder is only accurate
when working within the same millisecond.